THRIFT-4100: introducing a TSerializable trait in rust library Client: rs
* definition of a new TSerializable trait in the protocol module
* implementation of this trait for base types, list, maps, sets, Option, and Box types
* added a new subcrate "thrift-derive" to host a derive macro for TSerializable (cuz that's how it is done)
* modified the compiler to remove obsolete portions of the code that are now handled by the trait

I've tried to make the code generated by the derive macro be as close as possible to the code
formerly generated by the compiler, which makes things a bit ugly sometimes.
I had to slightly modify the return type of the read_from_in_protocol to a Result<Box<Self>> instead of Result<Self>.
This should not influence user's code, since these methods are called under the hood (maybe debatable :)).